update UI for Project Details and white back button

+++ IColab/Core/Data/Entities/Milestone.swift
@@ -10,7 +10,7 @@ import Foundation
 struct Milestone: Identifiable, Hashable {
     let id = UUID().uuidString
     var name : String
-    var nominal : Float
+    var nominal : Int
     var desc : String
     var endDate : Date
     var isAchieved : Bool
